chore(ci): close clippy debt across audio.terminus + playback.mpd + parity-harness
The v0.1.13 branch had accumulated seven clippy errors
across earlier commits on this branch, uncovered by a full
local gate run.

Bar-test of each closure. Every change strengthens an
invariant or retires complexity; none introduce parallel
truth paths.

audio.terminus:
- src/local_role.rs + src/transport_gate.rs: the
  `parse_from_state` projection fn and its only consumers
  (`local_role_subscriber`, `now_playing_subscriber`) are
  both gated behind the `alsa-substrate` feature. The
  consumers carried the gate; the projection fn didn't.
  Default-feature build saw an orphan public fn. Aligned
  the gates with
  `#[cfg(any(test, feature = "alsa-substrate"))]` so tests
  stay reachable in the default build and the
  alsa-substrate build picks up the production wiring.
  Same fix to the `use serde_json::Value;` import in both
  files.
- src/spectrum_subject.rs + src/lib.rs: doc comments
  started continuation lines with `+`, which CommonMark
  treats as a list-item-without-indent. Rephrased to prose
  ("X and Y", not "X / + Y") so the lint trips on intent,
  not formatting trick.
- src/spectrum_subject.rs: replaced `for i in 0..BIN_COUNT`
  used only as an index into `corr` with
  `corr.iter().take()`.
- src/lib.rs: dropped `as u32` on a value already typed
  `u32`.
- src/lib.rs + src/fft.rs: added missing doc comments to
  two pub items.

playback.mpd:
- src/idle_observer.rs: collapsed an inner `if !flag` guard
  into a match-guard on the enum arm so the precondition
  lives with the pattern, not buried two lines into the
  body.
- src/mpd/connection.rs: doc comment for `listallinfo`
  started a continuation line with `+`; rephrased to prose.
- src/mpd/types.rs `MpdLibraryEntry` +
  src/mpd/connection.rs `LibraryEntryBuilder`: applied
  `#[allow(clippy::large_enum_variant)]` with documented
  rationale. Entries are produced by the streaming
  `listallinfo` parser, projected to the wire envelope, and
  dropped — Boxing the File variant would add one heap
  allocation per parsed track (1000+ on a typical library
  walk) for zero correctness benefit. The lint's
  optimisation pressure does not apply to short-lived
  projection types.
- src/works.rs: replaced `let mut x = T::default();
  x.field = ...;` with a struct literal carrying
  `..Default::default()`.
- src/playback_supervisor/actor.rs: bundled
  `last_emitted_file` + `last_probed_file` (both
  `Option<String>` cross-call state, always read/updated
  together) into a new `FileEmissionTracker` struct.
  Reduced `emit_initial_report` /
  `emit_best_effort_report` from 8 args to 7; reduced
  `SupervisorTask` field count by one. Matches the
  codebase's existing bundling pattern (`EnvelopeState`,
  `BackoffState`, `PlaybackStateReport`). 95% of call-
  sites already used the two fields together; bundling
  makes the cohesion explicit.

parity-harness:
- src/queue.rs: replaced manual `let mut taken = 0;`
  counter + `if taken >= N { break; } taken += 1;` with
  `entries.into_iter().take(DISCOVERY_MAX_BREADTH)`. The
  iterator combinator is bounds-checked by construction;
  the manual counter could drift on future edits.

Verification:
- `cargo clean / fmt --check / clippy --workspace
  --all-targets --locked -- -D warnings / test --workspace
  --all-targets --locked / build --workspace --locked` —
  all clean.
- No semantic behaviour change in any change. No public-
  API break (the only signature changes are internal-fn-
  only).

+++ plugins/org.evoframework.audio.terminus/src/lib.rs
@@ -107,8 +107,8 @@ const PLUGIN_NAME: &str = "org.evoframework.audio.terminus";
 const TERMINUS_INPUT_PCM: &str = "hw:Loopback,1,7";
 
 /// Embedded in-process manifest. Production OOP path uses
-/// `manifest.oop.toml`; both share the same shelf + interaction
-/// + capabilities surface. Consumed by tests that assert the
+/// `manifest.oop.toml`; both share the same shelf, interaction,
+/// and capabilities surface. Consumed by tests that assert the
 /// describe() identity matches the manifest declaration.
 #[cfg(test)]
 const EMBEDDED_MANIFEST: &str = include_str!("../manifest.toml");
@@ -221,6 +221,9 @@ pub struct AudioTerminusPlugin {
 }
 
 impl AudioTerminusPlugin {
+    /// Construct an unloaded plugin. State is empty until
+    /// [`Plugin::load`] runs; calling any [`Respondent`] verb
+    /// before load returns [`PluginError::Permanent`].
     pub fn new() -> Self {
         Self {
             loaded: false,
@@ -544,7 +547,7 @@ mod tests {
         let desc = plugin.describe().await;
         let mf = parse_embedded_manifest();
         assert_eq!(desc.identity.name, mf.plugin.name);
-        assert_eq!(desc.identity.contract, mf.plugin.contract as u32);
+        assert_eq!(desc.identity.contract, mf.plugin.contract);
         assert!(!desc.runtime_capabilities.accepts_custody);
         // Every declared request type in the manifest must
         // appear in the runtime capabilities.
